chore(api): analysis: move are-puuids-allowed to shared provider
