+++ api/src/modules/analysis/providers/shared/__mocks__/scenario-planning-unit-geo.data.ts
